requirements.txt
appdirs==1.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
blinker==1.4
certifi==2017.4.17
chardet==3.0.4
click==6.7
configparser==3.5.0
decorator==4.0.11
entrypoints==0.2.3
enum34==1.1.6
Flask==0.9
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==0.16
Flask-WTF==0.8
functools32==3.2.3.post2
future==0.16.0
gyp==0.1
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
MarkupSafe==0.23
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
oauthlib==2.0.2
packaging==16.8
pandocfilters==1.4.1
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.6.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.1
python-twitter==3.3
pyzmq==16.0.2
requests==2.18.1
requests-oauthlib==0.8.0
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==0.7.9
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.21.1
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.8.3
WTForms==1.0.2
